feat: implement core functionality for subjects, courses, degrees, fields, grades, and professors
- Added SubjectController and associated routes for CRUD operations on subjects.
- Implemented CourseRepository and CourseService for course management.
- Created DegreeRepository and DegreeService for handling degrees.
- Developed FieldRepository and FieldService for field management.
- Introduced GradeRepository and GradeService for grade operations.
- Established Professor schema and service for managing professor data.
- Implemented authentication middleware and service for user login.
- Added validation schemas for courses, degrees, fields, grades, and subjects using Zod.
- Created API service for handling HTTP requests to the backend.
- Developed React Native screens for registering and managing professors.

+++ AppScholar-API/src/controllers/professor.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 import { professorService } from '../services/professor.service';
+import { AuthRequest } from '../schemas/auth.schema';
 
 class ProfessorController {
     
@@ -31,8 +32,15 @@ class ProfessorController {
         }
     }
 
-    public async updateProfessor(req: Request, res: Response, next: NextFunction): Promise<void> {
+    public async updateProfessor(req: AuthRequest, res: Response, next: NextFunction): Promise<void | Response<any>> {
         try {
+
+            const loggedUserRole = req.user.role;
+
+            if (loggedUserRole !== 'PROFESSOR' && loggedUserRole !== 'ADMIN') {
+                return res.status(403).json({ message: "Apenas professores e administradores podem editar informações." });
+            }
+            
             const userId = Number(req.params.userId);
             const updatedProfessor = await professorService.updateByUserId(userId, req.body);
             res.status(200).json(updatedProfessor);
@@ -41,8 +49,15 @@ class ProfessorController {
         }
     }
 
-    public async deleteProfessor(req: Request, res: Response, next: NextFunction): Promise<void> {
+    public async deleteProfessor(req: AuthRequest, res: Response, next: NextFunction): Promise<void | Response<any>> {
         try {
+
+            const loggedUserRole = req.user.role;
+
+            if (loggedUserRole !== 'ADMIN') {
+                return res.status(403).json({ message: "Apenas administradores podem excluir professores." });
+            }
+
             const userId = Number(req.params.userId);
             await professorService.deleteByUserId(userId);
             res.status(204).send();
